backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.2.0
celery[redis]>=5.3.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.5
openai>=1.30.0
anthropic>=0.28.0
sentry-sdk[fastapi]>=2.3.0
pydantic[email]>=2.7.0
python-multipart>=0.0.9
boto3>=1.34.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
aiosqlite>=0.20.0
redis>=5.0.0
httpx>=0.27.0
